@@ -1,7 +1,7 @@
 name: device_info_plus
 description: Flutter plugin providing detailed information about the device
   (make, model, etc.), and Android or iOS version the app is running on.
-version: 0.7.0
+version: 1.0.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
@@ -19,14 +19,17 @@ flutter:
         default_package: device_info_plus_web
       macos:
         default_package: device_info_plus_macos
+      windows:
+        default_package: device_info_plus_windows
 
 dependencies:
   flutter:
     sdk: flutter
-  device_info_plus_platform_interface: ^0.3.0
-  device_info_plus_linux: ^0.2.0
-  device_info_plus_macos: ^0.1.0
-  device_info_plus_web: ^0.3.0
+  device_info_plus_platform_interface: ^1.0.0-nullsafety.0
+  device_info_plus_linux: ^1.0.0-nullsafety.0
+  device_info_plus_macos: ^1.0.0-nullsafety.0
+  device_info_plus_web: ^1.0.0-nullsafety.0
+  device_info_plus_windows: ^1.0.0-nullsafety.0
 
 dev_dependencies:
   test: ^1.3.0
@@ -35,5 +38,5 @@ dev_dependencies:
   pedantic: ^1.9.2
 
 environment:
-  sdk: ">=2.1.0<3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
