@@ -1,15 +1,15 @@
 name: device_info_plus_linux
 description: Linux implementation of the device_info_plus plugin
-version: 0.2.0
+version: 1.0.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
-  device_info_plus_platform_interface: ^0.3.0
+  device_info_plus_platform_interface: ^1.0.0-nullsafety.0
   file: '>=5.2.1 <=6.0.0'
   flutter:
     sdk: flutter
