@@ -2,14 +2,14 @@ name: device_info_plus_macos
 description: Macos implementation of the device_info_plus plugin
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
-version: 0.1.0
+version: 1.0.0-nullsafety.0
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
-  device_info_plus_platform_interface: ^0.3.0
+  device_info_plus_platform_interface: ^1.0.0-nullsafety.0
   flutter:
     sdk: flutter
 
