@@ -1,6 +1,6 @@
 name: device_info_plus_web
 description: An implementation for the web platform of the Flutter `device_info_plus` plugin.
-version: 0.3.0
+version: 1.0.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
@@ -12,7 +12,7 @@ flutter:
         fileName: device_info_plus_web.dart
 
 dependencies:
-  device_info_plus_platform_interface: ^0.3.0
+  device_info_plus_platform_interface: ^1.0.0-nullsafety.0
   flutter_web_plugins:
     sdk: flutter
   flutter:
@@ -28,5 +28,5 @@ dev_dependencies:
   pedantic: ^1.9.2
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: '>=1.20.0'
