@@ -8,7 +8,7 @@ import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart' show kIsWeb, visibleForTesting;
 import 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart';
 import 'package:device_info_plus_linux/device_info_plus_linux.dart';
-
+import 'package:device_info_plus_windows/device_info_plus_windows.dart';
 export 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart'
     show
         AndroidBuildVersion,
@@ -17,6 +17,7 @@ export 'package:device_info_plus_platform_interface/device_info_plus_platform_in
         IosUtsname,
         LinuxDeviceInfo,
         MacOsDeviceInfo,
+        WindowsDeviceInfo,
         WebBrowserInfo;
 
 /// Provides device and operating system information.
@@ -34,20 +35,28 @@ class DeviceInfoPlugin {
   }
 
   static bool _disablePlatformOverride = false;
-  static DeviceInfoPlatform __platform;
+  static DeviceInfoPlatform? __platform;
 
   // This is to manually endorse the Linux plugin until automatic registration
   // of dart plugins is implemented.
   // See https://github.com/flutter/flutter/issues/52267 for more details.
   static DeviceInfoPlatform get _platform {
-    __platform ??= !kIsWeb && Platform.isLinux && !_disablePlatformOverride
-        ? DeviceInfoLinux()
-        : DeviceInfoPlatform.instance;
-    return __platform;
+    if (__platform == null) {
+      if (!_disablePlatformOverride && !kIsWeb) {
+        if (Platform.isLinux) {
+          __platform = DeviceInfoLinux();
+        } else if (Platform.isWindows) {
+          __platform = DeviceInfoWindows();
+        }
+      }
+      __platform ??= DeviceInfoPlatform.instance;
+    }
+
+    return __platform!;
   }
 
   /// This information does not change from call to call. Cache it.
-  AndroidDeviceInfo _cachedAndroidDeviceInfo;
+  AndroidDeviceInfo? _cachedAndroidDeviceInfo;
 
   /// Information derived from `android.os.Build`.
   ///
@@ -56,7 +65,7 @@ class DeviceInfoPlugin {
       _cachedAndroidDeviceInfo ??= await _platform.androidInfo();
 
   /// This information does not change from call to call. Cache it.
-  IosDeviceInfo _cachedIosDeviceInfo;
+  IosDeviceInfo? _cachedIosDeviceInfo;
 
   /// Information derived from `UIDevice`.
   ///
@@ -65,7 +74,7 @@ class DeviceInfoPlugin {
       _cachedIosDeviceInfo ??= await _platform.iosInfo();
 
   /// This information does not change from call to call. Cache it.
-  LinuxDeviceInfo _cachedLinuxDeviceInfo;
+  LinuxDeviceInfo? _cachedLinuxDeviceInfo;
 
   /// Information derived from `/etc/os-release`.
   ///
@@ -74,16 +83,22 @@ class DeviceInfoPlugin {
       _cachedLinuxDeviceInfo ??= await _platform.linuxInfo();
 
   /// This information does not change from call to call. Cache it.
-  WebBrowserInfo _cachedWebBrowserInfo;
+  WebBrowserInfo? _cachedWebBrowserInfo;
 
   /// Information derived from `Navigator`.
   Future<WebBrowserInfo> get webBrowserInfo async =>
       _cachedWebBrowserInfo ??= await _platform.webBrowserInfo();
 
   /// This information does not change from call to call. Cache it.
-  MacOsDeviceInfo _cachedMacosDeviceInfo;
+  MacOsDeviceInfo? _cachedMacosDeviceInfo;
 
   /// Returns device information for macos. Information sourced from Sysctl.
   Future<MacOsDeviceInfo> get macOsInfo async =>
       _cachedMacosDeviceInfo ??= await _platform.macosInfo();
+
+  WindowsDeviceInfo? _cachedWindowsDeviceInfo;
+
+  /// Returns device information for Windows.
+  Future<WindowsDeviceInfo> get windowsInfo async =>
+      _cachedWindowsDeviceInfo ??= await _platform.windowsInfo()!;
 }
