@@ -8,7 +8,7 @@ function toReferenceAPI(plugin) {
 
 module.exports = {
   main: {
-    "Getting Started": ["overview"],
-    Battery: ["battery/overview", toReferenceAPI("battery")],
+    "Getting Started": ["overview", "migration_guide"],
+    Battery: ["battery/overview","battery/usage", toReferenceAPI("battery")],
   },
 };
