@@ -4,6 +4,4 @@
 
 #include "generated_plugin_registrant.h"
 
-
-void fl_register_plugins(FlPluginRegistry* registry) {
-}
+void fl_register_plugins(FlPluginRegistry* registry) {}
