@@ -1,17 +1,17 @@
 #include "include/battery_plus_windows/battery_plus_windows_plugin.h"
-#include "include/battery_plus_windows/system_battery.h"
 
-#include <windows.h>
-
-#include <flutter/method_channel.h>
 #include <flutter/event_channel.h>
 #include <flutter/event_sink.h>
 #include <flutter/event_stream_handler.h>
+#include <flutter/method_channel.h>
 #include <flutter/plugin_registrar_windows.h>
 #include <flutter/standard_method_codec.h>
+#include <windows.h>
 
 #include <functional>
 #include <memory>
+
+#include "include/battery_plus_windows/system_battery.h"
 
 namespace {
 
@@ -21,7 +21,8 @@ typedef flutter::MethodCall<flutter::EncodableValue> FlMethodCall;
 typedef flutter::MethodResult<flutter::EncodableValue> FlMethodResult;
 typedef flutter::MethodChannel<flutter::EncodableValue> FlMethodChannel;
 typedef flutter::StreamHandler<flutter::EncodableValue> FlStreamHandler;
-typedef flutter::StreamHandlerError<flutter::EncodableValue> FlStreamHandlerError;
+typedef flutter::StreamHandlerError<flutter::EncodableValue>
+    FlStreamHandlerError;
 
 class BatteryPlusWindowsPlugin : public flutter::Plugin {
  public:
@@ -31,8 +32,8 @@ class BatteryPlusWindowsPlugin : public flutter::Plugin {
   ~BatteryPlusWindowsPlugin();
 
  private:
-  void HandleMethodCall(
-      const FlMethodCall &method_call, std::unique_ptr<FlMethodResult> result);
+  void HandleMethodCall(const FlMethodCall &method_call,
+                        std::unique_ptr<FlMethodResult> result);
 };
 
 class BatteryStatusStreamHandler : public FlStreamHandler {
@@ -43,11 +44,11 @@ class BatteryStatusStreamHandler : public FlStreamHandler {
   void AddStatusEvent(BatteryStatus status);
 
   std::unique_ptr<FlStreamHandlerError> OnListenInternal(
-      const flutter::EncodableValue* arguments,
-      std::unique_ptr<FlEventSink>&& events) override;
+      const flutter::EncodableValue *arguments,
+      std::unique_ptr<FlEventSink> &&events) override;
 
   std::unique_ptr<FlStreamHandlerError> OnCancelInternal(
-      const flutter::EncodableValue* arguments) override;
+      const flutter::EncodableValue *arguments) override;
 
  private:
   int _delegate = -1;
@@ -67,10 +68,9 @@ BatteryPlusWindowsPlugin::BatteryPlusWindowsPlugin(
       registrar->messenger(), "dev.fluttercommunity.plus/battery",
       &flutter::StandardMethodCodec::GetInstance());
 
-  methodChannel->SetMethodCallHandler(
-      [this](const auto &call, auto result) {
-        HandleMethodCall(call, std::move(result));
-      });
+  methodChannel->SetMethodCallHandler([this](const auto &call, auto result) {
+    HandleMethodCall(call, std::move(result));
+  });
 
   auto eventChannel = std::make_unique<FlEventChannel>(
       registrar->messenger(), "dev.fluttercommunity.plus/charging",
@@ -83,47 +83,52 @@ BatteryPlusWindowsPlugin::BatteryPlusWindowsPlugin(
 BatteryPlusWindowsPlugin::~BatteryPlusWindowsPlugin() {}
 
 BatteryStatusStreamHandler::BatteryStatusStreamHandler(
-  flutter::PluginRegistrarWindows *registrar) : _registrar(registrar) {}
+    flutter::PluginRegistrarWindows *registrar)
+    : _registrar(registrar) {}
 
 void BatteryStatusStreamHandler::AddStatusEvent(BatteryStatus status) {
-    if (status != BatteryStatus::Error) {
-      _events->Success(_battery.GetStatusString());
-    } else {
-      _events->Error(std::to_string(_battery.GetError()), _battery.GetErrorString());
-    }
+  if (status != BatteryStatus::Error) {
+    _events->Success(_battery.GetStatusString());
+  } else {
+    _events->Error(std::to_string(_battery.GetError()),
+                   _battery.GetErrorString());
+  }
 }
 
-std::unique_ptr<FlStreamHandlerError> BatteryStatusStreamHandler::OnListenInternal(
-    const flutter::EncodableValue* arguments,
-    std::unique_ptr<FlEventSink>&& events) {
+std::unique_ptr<FlStreamHandlerError>
+BatteryStatusStreamHandler::OnListenInternal(
+    const flutter::EncodableValue *arguments,
+    std::unique_ptr<FlEventSink> &&events) {
   _events = std::move(events);
 
   HWND hwnd = _registrar->GetView()->GetNativeWindow();
-  BatteryStatusCallback callback =
-      std::bind(&BatteryStatusStreamHandler::AddStatusEvent,
-                this, std::placeholders::_1);
+  BatteryStatusCallback callback = std::bind(
+      &BatteryStatusStreamHandler::AddStatusEvent, this, std::placeholders::_1);
 
   if (!_battery.StartListen(hwnd, callback)) {
     return std::make_unique<FlStreamHandlerError>(
-      std::to_string(_battery.GetError()), _battery.GetErrorString(), nullptr);
+        std::to_string(_battery.GetError()), _battery.GetErrorString(),
+        nullptr);
   }
 
   _delegate = _registrar->RegisterTopLevelWindowProcDelegate(
-    [this](HWND hwnd, UINT message, WPARAM wparam, LPARAM lparam) {
-      _battery.ProcessMsg(hwnd, message, wparam, lparam);
-      return std::nullopt;
-    });
+      [this](HWND hwnd, UINT message, WPARAM wparam, LPARAM lparam) {
+        _battery.ProcessMsg(hwnd, message, wparam, lparam);
+        return std::nullopt;
+      });
 
   AddStatusEvent(_battery.GetStatus());
   return nullptr;
 }
 
-std::unique_ptr<FlStreamHandlerError> BatteryStatusStreamHandler::OnCancelInternal(
-    const flutter::EncodableValue* arguments) {
+std::unique_ptr<FlStreamHandlerError>
+BatteryStatusStreamHandler::OnCancelInternal(
+    const flutter::EncodableValue *arguments) {
   _registrar->UnregisterTopLevelWindowProcDelegate(_delegate);
   if (!_battery.StopListen()) {
     return std::make_unique<FlStreamHandlerError>(
-      std::to_string(_battery.GetError()), _battery.GetErrorString(), nullptr);
+        std::to_string(_battery.GetError()), _battery.GetErrorString(),
+        nullptr);
   }
   _delegate = -1;
   _events.reset();
@@ -131,15 +136,15 @@ std::unique_ptr<FlStreamHandlerError> BatteryStatusStreamHandler::OnCancelIntern
 }
 
 void BatteryPlusWindowsPlugin::HandleMethodCall(
-    const FlMethodCall &method_call,
-    std::unique_ptr<FlMethodResult> result) {
+    const FlMethodCall &method_call, std::unique_ptr<FlMethodResult> result) {
   if (method_call.method_name().compare("getBatteryLevel") == 0) {
     SystemBattery battery;
     int level = battery.GetLevel();
     if (level >= 0) {
       result->Success(flutter::EncodableValue(level));
     } else {
-      result->Error(std::to_string(battery.GetError()), battery.GetErrorString());
+      result->Error(std::to_string(battery.GetError()),
+                    battery.GetErrorString());
     }
   } else {
     result->NotImplemented();
