@@ -5,7 +5,7 @@
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:connectivity_plus_platform_interface/src/method_channel_connectivity.dart';
+import 'package:connectivity_plus_platform_interface/method_channel_connectivity.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
