@@ -37,7 +37,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> initPlatformState() async {
-    Map<String, dynamic> deviceData;
+    var deviceData = <String, dynamic>{};
 
     try {
       if (kIsWeb) {
@@ -52,6 +52,9 @@ class _MyAppState extends State<MyApp> {
           deviceData = _readLinuxDeviceInfo(await deviceInfoPlugin.linuxInfo);
         } else if (Platform.isMacOS) {
           deviceData = _readMacOsDeviceInfo(await deviceInfoPlugin.macOsInfo);
+        } else if (Platform.isWindows) {
+          deviceData =
+              _readWindowsDeviceInfo(await deviceInfoPlugin.windowsInfo);
         }
       }
     } on PlatformException {
@@ -166,6 +169,14 @@ class _MyAppState extends State<MyApp> {
     };
   }
 
+  Map<String, dynamic> _readWindowsDeviceInfo(WindowsDeviceInfo data) {
+    return <String, dynamic>{
+      'numberOfCores': data.numberOfCores,
+      'computerName': data.computerName,
+      'systemMemoryInMegabytes': data.systemMemoryInMegabytes,
+    };
+  }
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -182,7 +193,9 @@ class _MyAppState extends State<MyApp> {
                             ? 'Linux Device Info'
                             : Platform.isMacOS
                                 ? 'MacOS Device Info'
-                                : '',
+                                : Platform.isWindows
+                                    ? 'Windows Device Info'
+                                    : '',
           ),
         ),
         body: ListView(
