@@ -1,15 +1,16 @@
 #ifndef FLUTTER_PLUGIN_NETWORK_INFO_PLUS_NETWORK_INFO_H_
 #define FLUTTER_PLUGIN_NETWORK_INFO_PLUS_NETWORK_INFO_H_
 
-#include <winsock2.h>
 #include <windows.h>
 #include <winerror.h>
+#include <winsock2.h>
 #include <wlanapi.h>
 
 #include <functional>
 #include <string>
 
-typedef std::function<std::string(LPGUID, PWLAN_CONNECTION_ATTRIBUTES)> WlanQuery;
+typedef std::function<std::string(LPGUID, PWLAN_CONNECTION_ATTRIBUTES)>
+    WlanQuery;
 
 class NetworkInfo {
  public:
