@@ -4,6 +4,4 @@
 
 #include "generated_plugin_registrant.h"
 
-
-void RegisterPlugins(flutter::PluginRegistry* registry) {
-}
+void RegisterPlugins(flutter::PluginRegistry* registry) {}
